Bug 16502: Table koha_plugin_com_bywatersolutions_kitchensink_mytable not always dropped after running Plugin.t
If you run Plugin.t, the above table will still be present (when you
did not enable UseKohaPlugins). This would trigger a warning when
running the test a second time.

Why? The uninstall call does its work not completely due to a small
inconsistency in Koha::Plugins::Handler::delete when calling run
without the enable_plugins parameter.

This patch resolves that inconsistency and also removes an unneeded skip
in Plugin.t in case the KitchenSink module already exists.
Note: This is a small fix. But I wonder if the Handler routines run and
delete should not have been implemented in Koha::Plugins::Base.
Also note that plugins/plugins-uninstall.pl will not be affacted by this
change, since it checks whether the pref is enabled before calling the
delete method.

Test plan:
[1] Do not yet install this patch.
[2] Verify that plugins are enabled in koha-conf.xml.
[3] Disable UseKohaPlugins in System Preferences!
[4] Run t/db_dependent/Plugins.t.
[5] Verify that table koha_plugin_com_bywatersolutions_kitchensink_mytable
    still exists. (It should have been deleted.) Remove it manually.
[6] Apply this patch.
[7] Run the test again.
[8] Verify that the table does not exist.
[9] Run the test again (without warnings).



(cherry picked from commit 4263ac2b92737024d8d620a751babf72b904b73a)
(cherry picked from commit 4ebe7b489c9798d2456bd3de1d95ec6e027b2b21)
(cherry picked from commit 752e31425efa34fa6f21446fb18cf34ba31fc441)

+++ Koha/Plugins/Handler.pm
@@ -78,11 +78,18 @@ Deletes a plugin
 
 sub delete {
     my ( $class, $args ) = @_;
+
+    return unless ( C4::Context->config("enable_plugins") || $args->{'enable_plugins'} );
+
     my $plugin_class = $args->{'class'};
     my $plugin_dir   = C4::Context->config("pluginsdir");
     my $plugin_path  = "$plugin_dir/" . join( '/', split( '::', $args->{'class'} ) );
 
-    Koha::Plugins::Handler->run( { class => $plugin_class, method => 'uninstall' } );
+    Koha::Plugins::Handler->run({
+        class          => $plugin_class,
+        method         => 'uninstall',
+        enable_plugins => $args->{enable_plugins},
+    });
 
     C4::Context->dbh->do( "DELETE FROM plugin_data WHERE plugin_class = ?", undef, ($plugin_class) );
